refactored components for mobile into separate components
